+++ src/DpressWebApp.php
@@ -31,6 +31,7 @@ use Dynart\Dpress\Controller\Admin\UserAdminController;
 use Dynart\Dpress\Controller\AuthController;
 use Dynart\Dpress\Controller\ContentController;
 use Dynart\Dpress\Controller\FeedController;
+use Dynart\Dpress\Controller\SitemapController;
 use Dynart\Dpress\Controller\HomeController;
 use Dynart\Dpress\Controller\MediaController;
 use Dynart\Dpress\Controller\PageController;
@@ -68,6 +69,7 @@ class DpressWebApp extends WebApp {
         ProfileController::class,
         ContentController::class,
         FeedController::class,
+        SitemapController::class,
         MediaController::class,
         AssetController::class,
         DashboardController::class,